chore: fix all clippy warnings and fmt
- Add missing Port imports in test modules
- Fix ControllerCollection::new arg count in rom_test
- Fix read_port signatures to use &dyn Port
- Run cargo fmt --all

+++ nes/core/src/cpu/memory.rs
@@ -48,7 +48,7 @@ impl Memory {
                 ppu.read_register(address, &mut ppu_cartridge, interrupt)
             }
             0x4015 => apu.read_register(address, interrupt),
-            0x4016 | 0x4017 => hub.read_port(&NES_PORTS[address as usize & 1]),
+            0x4016 | 0x4017 => hub.read_port(&NES_PORTS[address & 1]),
             0x4000..=0x4014 | 0x4018..=0x401F => OpenBusReadResult::new(0, 0), // TODO: I/O registers
             0x4020..=0x5FFF => cartridge.read(address),
             0x6000..=0xFFFF => cartridge.read(address),
